fix: make sure only terminte inactive worker

+++ src/main_thread/api/public_api.ts
@@ -490,10 +490,22 @@ class Player extends EventEmitter<IPublicAPIEvent> {
         );
       }
 
-      // if the user already attach worker before
+      // check if the user already attach worker before
       // terminate the previous worker to release the resources
       if (this._priv_worker !== null) {
-        this._priv_worker.terminate();
+        if (this.__priv_isPlaying()) {
+          log.warn(
+            "API: Cannot attach a new worker while a content is playing, please stop the player first.",
+          );
+          return rej(
+            new WorkerInitializationError(
+              "SETUP_ERROR",
+              "Cannot attach a new worker while a content is playing",
+            ),
+          );
+        } else {
+          this._priv_worker.terminate();
+        }
       }
 
       if (typeof workerSettings.workerUrl === "string") {
@@ -838,6 +850,13 @@ class Player extends EventEmitter<IPublicAPIEvent> {
       );
     }
     return renderThumbnail(this._priv_contentInfos, options);
+  }
+
+  /**
+   * Get the player is currently playing
+   */
+  private __priv_isPlaying(): boolean {
+    return !["ENDED", "STOPPED"].includes(this.state);
   }
 
   /**
